requirements.txt
Django==2.2.10
django-filter==2.4.0
djangorestframework==3.12.2
